requirements.txt
flask==3.1.0
Pyjwt==2.10.1
flask-cors==5.0.1
python-dotenv==1.1.0
requests==2.32.3
youtube-transcript-api==1.0.3
langchain-google-genai==2.1.3
apscheduler==3.11.0
feedparser==6.0.11
newspaper3k==0.2.8
lxml_html_clean==0.4.2
sqlalchemy==2.0.40
psycopg2[binary]==2.9.10
